feat(web): W2 — wasm platform shims + prototype-drift reconciliation
Gate the native crash-recovery signal layer (main_pc.c) and the stall
watchdog thread/posix_spawn (stall_watchdog.c) behind a 3rd __EMSCRIPTEN__
fork — neither has a single-threaded-wasm runtime backing.

Reconcile the 6 wasm-ld function-signature-mismatch warnings W1.5 recorded
(hard indirect-call traps under wasm, latent prototype-drift UB on native):
 - randomSetSeed: random.h u32 -> u64 (definition is truth; g_randomSeed is
   u64 / 64-bit asm). Shared fix, byte-identical on native.
 - texLoadFromDisplayList / getRoomPositionScaledByIndex: implicit int-return
   decls -> add the real void prototype (image.h) / include (lightfixture.c).
   Shared, return unused, behaviour-neutral.
 - acos/asin: game fixed-point trig collides with SDL libm only under wasm's
   single-module static link. Rename to ge007_fixed_* for wasm only; native
   keeps retail names and the shim math.h untouched.
 - sub_GAME_7F0B4F9C: stan.c's deliberately-"incorrect" s32 asm-match decl
   traps against bg.c's f32 definition. wasm-only true prototype; the caller
   (sub_GAME_7F0B2FE0) is dead code, so behaviour-neutral.

Native gates green: ge007 clean build; ctest
port_train_faithful_aperture_regression + port_rom_reject_guard 2/2; all 7
fidelity tapes byte-exact. wasm: zero wasm-ld warnings, clean link.

+++ src/random.h
@@ -2,7 +2,12 @@
 #define _RANDOM_H_
 #include <ultra64.h>
 
-void randomSetSeed(u32 param_1);
+/* Definition (random.c: g_randomSeed is u64; the retail asm does a 64-bit
+ * `daddiu`/`sd`) takes a u64. The prior u32 prototype was prototype drift:
+ * benign on native (callers pass osGetCount()/a literal, both zero-extend to
+ * the same value) but a hard function-signature trap under wasm's indirect-
+ * call type checks. The definition is truth. */
+void randomSetSeed(u64 param_1);
 u32 randomGetNext(void);
 u32 randomGetNextFrom(u64 *param_1);
 #ifdef NATIVE_PORT
